requirements.txt
Flask==2.0.2
keras==2.6.0
matplotlib==3.3.4
numpy==1.19.5
pandas==1.1.5
Pillow==8.4.0
scikit_learn==1.0.1
scipy==1.5.4
seaborn==0.11.2
streamlit==1.2.0
tensorflow==2.6.0
utils==1.0.1
WTForms==2.3.3
